SPARKC-590: DataSourceV2 Read Support
In this commit we refactor out a large deal of the CassandraTableScanRDD
code into pure functions held in ScanHelper. This allows us to reuse
this code in the new DSV2 Read path which no longer allows directly
delegating to an RDD. In DSV2 a new RDD is constructed using
ReaderFactories which generatate their own readers. We have also added
basic code around passing C* Clustering information which allows for
Spark to take advantage of underlying C* partitioning during operations.