SPARKC-590: Count Support
Previously we split responsibilites between ScanBuilder and Scan for
doing some of our analysis of the query to be built. I realized it makes
more logical sense to move this all as high as possible in the call
stack. This makes it easier to apply count optimizations without passing
through lots of "magic" empty arrays.

After doing so Count is now supported by a second PartitionReaderFactory
which acts like the Normal Scan Reader but instead of generating 1 for 1
rows out of Cassandra, it preforms a count, then generates empty rows
for the count result.