SPARKC-590: Get TTL Metadata Working Again
Only a slight refactoring was required to make the Metadata functions
work again. Basically all references to "relation" are changed to
DSV2Relation and we can modify the table there.

One limitation that is introduced is that the option process of
specifying ttls "ttl.colname = ttlName" is no longer valid. Because of
some of the new verification for output, we need to modify the plan
before the scan to have the correct number of ouputs. This means that
only via our Rule can we handle new columns. All behaviors with the
functions are as before so you can still select ("ttl(col)").