SPARKC-590: Fix Various Test Issues
One recurrent probelm in many of our SQL based tests is that determining
what SparkSession calls "spark.read.datasource" is no longer easy to
tell from inside the default source. Using SparkSession.active seems
like the correct thing to do but our test code plays rather fast and
loose with it's session references. Here we try to make explicit which
session is the active on in tests which require manipulation of the
spark session.

In addition a small bug with the type conversion of java.time.LocalTime
was corrected and a incorrect "contains" usage inside the
joinWithCassandraTable code.